+++ packages/core/src/components/cv-number-input/cv-number-input.vue
@@ -1,7 +1,6 @@
 <template>
   <cv-wrapper :tag-type="formItem ? 'div' : ''" class="cv-number-input" :class="`${carbonPrefix}--form-item`">
     <div
-      data-numberinput
       :class="[
         `${carbonPrefix}--number`,
         {
@@ -14,20 +13,12 @@
       :data-invalid="isInvalid"
     >
       <label :for="uid" :class="`${carbonPrefix}--label`">{{ label }}</label>
-      <div :class="`${carbonPrefix}--form__helper-text`" v-if="isHelper && mobile">
-        <slot name="helper-text">{{ helperText }}</slot>
-      </div>
-      <div :class="`${carbonPrefix}--number__input-wrapper`">
-        <button
-          v-if="mobile"
-          :class="`${carbonPrefix}--number__control-btn down-icon`"
-          @click="doDown"
-          type="button"
-          :aria-label="ariaLabelForDownButton"
-          :disabled="disabled"
-        >
-          <Subtract16 />
-        </button>
+      <div
+        :class="[
+          `${carbonPrefix}--number__input-wrapper`,
+          { [`${carbonPrefix}--number__input-wrapper--warning`]: isWarn },
+        ]"
+      >
         <input
           :data-invalid="isInvalid"
           :id="uid"
@@ -41,8 +32,9 @@
           :max="max"
           ref="input"
         />
-        <WarningFilled16 v-if="isInvalid && !mobile" :class="`${carbonPrefix}--number__invalid`" />
-        <div :class="`${carbonPrefix}--number__controls`" v-if="!mobile">
+        <WarningFilled16 v-if="isInvalid" :class="iconClasses" />
+        <WarningAltFilled16 v-if="isWarn && !isInvalid" :class="iconClasses" />
+        <div :class="`${carbonPrefix}--number__controls`">
           <button
             :class="`${carbonPrefix}--number__control-btn down-icon`"
             @click="doDown"
@@ -68,21 +60,14 @@
           </button>
           <div :class="`${carbonPrefix}--number__rule-divider`" />
         </div>
-        <button
-          v-else
-          :class="`${carbonPrefix}--number__control-btn up-icon`"
-          @click="doUp"
-          type="button"
-          :aria-label="ariaLabelForUpButton"
-          :disabled="disabled"
-        >
-          <Add16 />
-        </button>
       </div>
       <div :class="`${carbonPrefix}--form-requirement`" v-if="isInvalid">
         <slot name="invalid-message">{{ invalidMessage }}</slot>
       </div>
-      <div :class="`${carbonPrefix}--form__helper-text`" v-if="!isInvalid && isHelper && !mobile">
+      <div :class="`${carbonPrefix}--form-requirement`" v-if="isWarn && !isInvalid">
+        <slot name="warn-text">{{ warnText }}</slot>
+      </div>
+      <div :class="`${carbonPrefix}--form__helper-text`" v-if="!(isInvalid || isWarn) && isHelper">
         <slot name="helper-text">{{ helperText }}</slot>
       </div>
     </div>
@@ -94,12 +79,13 @@ import { uidMixin, themeMixin, carbonPrefixMixin, methodsMixin } from '../../mix
 import Add16 from '@carbon/icons-vue/es/add/16';
 import Subtract16 from '@carbon/icons-vue/es/subtract/16';
 import WarningFilled16 from '@carbon/icons-vue/es/warning--filled/16';
+import WarningAltFilled16 from '@carbon/icons-vue/es/warning--alt--filled/16';
 import CvWrapper from '../cv-wrapper/_cv-wrapper';
 
 export default {
   name: 'CvNumberInput',
   mixins: [uidMixin, themeMixin, carbonPrefixMixin, methodsMixin({ input: ['blur', 'focus'] })],
-  components: { Add16, Subtract16, WarningFilled16, CvWrapper },
+  components: { Add16, Subtract16, WarningFilled16, WarningAltFilled16, CvWrapper },
   inheritAttrs: false,
   props: {
     disabled: Boolean,
@@ -111,6 +97,7 @@ export default {
     invalid: /* deprecate */ {
       type: Boolean,
       default: undefined,
+      deprecated: true,
       validator(val) {
         if (val !== undefined && process.env.NODE_ENV === 'development') {
           console.warn('CvNumberInput: invalid prop deprecated in favour of invalidMessage');
@@ -123,13 +110,25 @@ export default {
     min: { type: [String, Number], default: undefined },
     max: { type: [String, Number], default: undefined },
     step: { type: [String, Number], default: undefined },
-    mobile: Boolean,
+    mobile: /* deprecate */ {
+      type: Boolean,
+      default: undefined,
+      deprecated: true,
+      validator(val) {
+        if (val !== undefined) {
+          console.warn('CvNumberInput: mobile prop deprecated as no longer different to standard number input.');
+        }
+        return true;
+      },
+    },
+    warnText: { type: String, default: undefined },
   },
   data() {
     return {
       internalValue: 0,
       isHelper: false,
       isInvalid: false,
+      isWarn: false,
     };
   },
   mounted() {
@@ -159,6 +158,14 @@ export default {
         input: ev => this.onInput(ev.target.value),
       });
     },
+    iconClasses() {
+      const classes = [];
+      (this.isInvalid || this.isWarn) && classes.push(`${this.carbonPrefix}--number__invalid`);
+      this.isWarn && !this.isInvalid && classes.push(`${this.carbonPrefix}--number__invalid--warning`);
+      this.readOnly && classes.push(`${this.carbonPrefix}--number__readonly-icon`);
+
+      return classes.join(' ');
+    },
   },
   methods: {
     onInput(val) {
@@ -168,6 +175,7 @@ export default {
     checkSlots() {
       // NOTE: this.$slots is not reactive so needs to be managed on updated
       this.isInvalid = !!(this.$slots['invalid-message'] || (this.invalidMessage && this.invalidMessage.length));
+      this.isWarn = !!(this.$slots['warn-text'] || (this.warnText && this.warnText.length));
       this.isHelper = !!(this.$slots['helper-text'] || (this.helperText && this.helperText.length));
     },
     doUp() {